+++ app/installer/app/components/package-manager.js
@@ -23,11 +23,11 @@ module.exports = {
         load: function () {
             this.$set('status', 'loading');
 
-            this.queryUpdates(this.packages, function (res) {
+            this.queryUpdates(this.packages).then(function (res) {
                 var data = res.data;
                 this.$set('updates', data.packages.length ? _.indexBy(data.packages, 'name') : null);
                 this.$set('status', '');
-            }).error(function () {
+            }, function () {
                 this.$set('status', 'error');
             });
         },
@@ -113,7 +113,8 @@ module.exports = {
 
     components: {
 
-        'package-upload': require('./package-upload.vue')
+        'package-upload': require('./package-upload.vue'),
+        'package-details': require('./package-details.vue')
 
     }