+++ app/installer/src/Helper/Composer.php
@@ -6,8 +6,10 @@ use Composer\Installer;
 use Composer\IO\ConsoleIO;
 use Composer\Json\JsonFile;
 use Composer\Package\Locker;
+use Composer\Package\Package;
 use Composer\Repository\CompositeRepository;
 use Composer\Repository\InstalledFilesystemRepository;
+use Composer\Semver\VersionParser;
 use Symfony\Component\Console\Output\OutputInterface;
 
 
@@ -59,7 +61,16 @@ class Composer
     {
         $this->addPackages($install);
 
-        $this->composerUpdate(array_keys($install));
+        $refresh = [];
+        $versionParser = new VersionParser();
+        foreach ($install as $name => $version) {
+            try {
+                $normalized = $versionParser->normalize($version);
+                $refresh[] = new Package($name, $normalized, $version);
+            } catch (\UnexpectedValueException $e) {}
+        }
+
+        $this->composerUpdate(array_keys($install), $refresh);
         $this->writeConfig();
     }
 
@@ -69,7 +80,7 @@ class Composer
      */
     public function uninstall($uninstall)
     {
-        $uninstall = (array)$uninstall;
+        $uninstall = (array) $uninstall;
 
         $this->removePackages($uninstall);
 
@@ -99,15 +110,22 @@ class Composer
      * Runs Composer Update command.
      *
      * @param  array|bool $updates
+     * @param array $refresh
      * @return bool
      */
-    protected function composerUpdate($updates = false)
+    protected function composerUpdate($updates = false, $refresh = [])
     {
         $installed = new JsonFile($this->paths['path.vendor'] . '/composer/installed.json');
         $internal = new CompositeRepository([]);
         $internal->addRepository(new InstalledFilesystemRepository($installed));
 
         $composer = $this->getComposer();
+        $composer->getDownloadManager()->setOutputProgress(false);
+
+        $local = $composer->getRepositoryManager()->getLocalRepository();
+        foreach ($refresh as $package) {
+            $local->removePackage($package);
+        }
 
         $installer = Installer::create($this->getIO(), $composer)
             ->setAdditionalInstalledRepository($internal)
@@ -116,7 +134,7 @@ class Composer
             ->setUpdate(true);
 
         if ($updates) {
-            $installer->setUpdateWhitelist($updates);
+            $installer->setUpdateWhitelist($updates)->setWhitelistDependencies();
         }
 
         $installer->run();
@@ -130,7 +148,7 @@ class Composer
     protected function getComposer()
     {
         $config = $this->blueprint;
-        $config['config'] = ['vendor-dir' => $this->paths['path.packages']];
+        $config['config'] = ['vendor-dir' => $this->paths['path.packages'], 'cache-files-ttl' => 0];
         $config['require'] = $this->packages;
 
         // set memory limit, if < 512M
@@ -210,7 +228,7 @@ class Composer
     protected function memoryInBytes($value)
     {
         $unit = strtolower(substr($value, -1, 1));
-        $value = (int)$value;
+        $value = (int) $value;
 
         switch ($unit) {
             case 'g':