+++ app/modules/debug/app/components/auth.vue
@@ -1,26 +1,26 @@
 <template>
 
-    <a title="User"><span class="pf-icon pf-icon-auth" :class="{'pf-parent': user}"></span>{{ label }}</a>
+    <a title="User"><span class="pf-icon pf-icon-auth" :class="{'pf-parent': data.user}"></span>{{ label }}</a>
 
-    <div class="pf-dropdown" v-show="user">
+    <div class="pf-dropdown" v-show="data.user">
 
         <table class="pf-table pf-table-dropdown">
             <tbody>
                 <tr>
                     <td>Username</td>
-                    <td>{{ user }}</td>
+                    <td>{{ data.user }}</td>
                 </tr>
                 <tr>
                     <td>Roles</td>
-                    <td>{{ roles | json }}</td>
+                    <td>{{ data.roles | json }}</td>
                 </tr>
                 <tr>
                     <td>Authenticated</td>
-                    <td>{{ authenticated ? 'yes' : 'no' }}</td>
+                    <td>{{ data.authenticated ? 'yes' : 'no' }}</td>
                 </tr>
                 <tr>
                     <td>Class</td>
-                    <td>{{ user_class }}</td>
+                    <td>{{ data.user_class }}</td>
                 </tr>
             </tbody>
         </table>
@@ -39,19 +39,15 @@
 
         props: ['data'],
 
-        data: function () {
-            return this.data;
-        },
-
         computed: {
 
             label: function () {
 
-                if (this.user) {
-                    return this.user;
+                if (this.data.user) {
+                    return this.data.user;
                 }
 
-                return this.enabled ? 'You are not authenticated.' : 'Authentication is disabled.';
+                return this.data.enabled ? 'You are not authenticated.' : 'Authentication is disabled.';
             }
 
         }