+++ app/modules/session/src/Csrf/Event/CsrfListener.php
@@ -3,6 +3,7 @@
 namespace Pagekit\Session\Csrf\Event;
 
 use Pagekit\Event\EventSubscriberInterface;
+use Pagekit\Session\Csrf\Exception\CsrfException;
 use Pagekit\Session\Csrf\Provider\CsrfProviderInterface;
 
 class CsrfListener implements EventSubscriberInterface
@@ -26,14 +27,14 @@ class CsrfListener implements EventSubscriberInterface
      * Checks for the CSRF token and throws 401 exception if invalid.
      *
      * @param  $event
-     * @throws \RuntimeException
+     * @throws UnauthorizedException
      */
     public function onRequest($event, $request)
     {
         $this->provider->setToken($request->get('_csrf', $request->headers->get('X-XSRF-TOKEN')));
 
-        if ($csrf = $request->attributes->get('_request[csrf]', false, true) and !$this->provider->validate()) {
-            throw new \RuntimeException('Invalid CSRF token.');
+        if ($request->attributes->get('_request[csrf]', false, true) && !$this->provider->validate()) {
+            throw new CsrfException('Invalid CSRF token.');
         }
     }