+++ app/system/modules/theme/views/login.php
@@ -41,7 +41,7 @@
                     </div>
 
                     <ul class="uk-list uk-contrast">
-                        <li><label class="uk-form"><input type="checkbox" name="<?= $remember_me_param ?>"> <?= __('Remember Me') ?></label></li>
+                        <li><label class="uk-form"><input type="checkbox" name="remember_me"> <?= __('Remember Me') ?></label></li>
                         <li class="uk-margin-small-top"> <?= __('Forgot Password?') ?> <a class="uk-link" data-uk-toggle="{ target: '.js-toggle' }"><?= __('Request Password') ?></a></li>
                     </ul>